requirements.txt
requests
IMDbPY
tgcrypto
pyrogram
pyyaml
